backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
Marshmallow==3.20.1
flask-cors==4.0.0
Click==8.1.7
pytest==7.4.3
python-dotenv==1.0.0